fixed a bunch of python 3 errors, division between integers will produce a float in python 3, I must be better somehow

+++ stencil_code/backend/ocl_boundary_copier.py
@@ -87,7 +87,7 @@ class OclBoundaryCopier(object):
         :return:
         """
         return [
-            size if size % self.local_size[dim] == 0 else ((size / self.local_size[dim]) + 1) * self.local_size[dim]
+            size if size % self.local_size[dim] == 0 else (int((size / self.local_size[dim]) + 1) * self.local_size[dim])
             for dim, size in enumerate(self.global_size)
         ]
